Refactor: spinner 제거 코드 리팩토링
