+++ cypress/integration/racing/racing.spec.js
@@ -84,7 +84,7 @@ describe('레이싱 테스트', () => {
     })
 
     it('경기가 종료되면 스핀이 사라진다.', () => {
-
+      
     })
   })
 });
