+++ src/js/View/Render/renderCountForm.js
@@ -1,5 +1,5 @@
 import { selector } from "../../util/consts.js";
-import { render, Component } from "../index.js"
+import { render, Component } from "../render.js"
 import Fieldset from "../Component/Form/Fieldset.js";
 
 const renderCountForm = () => {
