+++ src/js/View/Render/renderRacingArena.js
@@ -1,5 +1,5 @@
 import { selector } from "../../util/consts.js";
-import { render, Component } from "../index.js"
+import { render, Component } from "../render.js"
 
 import RacingGameArena from "../Component/Game/RacingGame.js";
 import Track from "../Component/Game/Track.js";
